+++ src/app/providers/ThemeProviders/index.ts
@@ -1,0 +1,4 @@
+import ThemeProvider from "./ui/ThemeProvider";
+import { ThemeContext, LOCAL_STORAGE_THEME_KEY, Theme } from "./lib/ThemeContext";
+
+export { ThemeProvider, ThemeContext, LOCAL_STORAGE_THEME_KEY, Theme };
