+++ src/app/providers/ThemeProviders/lib/ToggleTheme.ts
@@ -1,5 +1,9 @@
 import { useContext } from "react";
-import { Theme, ThemeContext, LOCAL_STORAGE_THEME_KEY } from "./ThemeContext";
+import {
+  ThemeContext,
+  LOCAL_STORAGE_THEME_KEY,
+  Theme,
+} from "app/providers/ThemeProviders";
 
 interface UseThemeResult {
   toggleTheme: () => void;
