+++ src/app/providers/ThemeProviders/ui/ThemeProvider.tsx
@@ -1,5 +1,9 @@
 import { FC, useMemo, useState } from "react";
-import { ThemeContext, Theme, LOCAL_STORAGE_THEME_KEY } from "./ThemeContext";
+import {
+  ThemeContext,
+  Theme,
+  LOCAL_STORAGE_THEME_KEY,
+} from "../lib/ThemeContext";
 
 const defaultTheme =
   (localStorage.getItem(LOCAL_STORAGE_THEME_KEY) as Theme) || Theme.LIGHT;
