+++ src/pages/About/index.tsx
@@ -1,5 +1,1 @@
-const AboutPage = () => {
-  return <div>AboutPage</div>;
-};
-
-export default AboutPage;
+export { AboutPageAsync as AboutPage } from "./ui/AboutPage.async";
