+++ src/pages/About/ui/AboutPage.tsx
@@ -1,0 +1,5 @@
+const AboutPage = () => {
+  return <div>AboutPage</div>;
+};
+
+export default AboutPage;
