+++ src/pages/Main/ui/MainPage.async.tsx
@@ -1,0 +1,10 @@
+import { lazy } from "react";
+
+export const MainPageAsync = lazy(
+  () =>
+    new Promise((resolve) => {
+      // @ts-ignore
+      // JUST FOR COURSE!
+      setTimeout(() => resolve(import("./MainPage")), 500);
+    })
+);
