+++ src/pages/Main/ui/MainPage.tsx
@@ -1,0 +1,5 @@
+const MainPage = () => {
+  return <div>MainPage</div>;
+};
+
+export default MainPage;
