+++ tsconfig.json
@@ -13,6 +13,11 @@
     "allowJs": true,
     "baseUrl": ".",
     "moduleResolution": "node",
+    "paths": {
+      "*": [
+        "./src/*"
+      ]
+    },
     // esModuleInterop позволяет работать с пакетами, которые используют
     // common js как с обычными пакета с помощью import (require() module.export) = common js
     // import Moment from 'moment';
